Update - 인증 정보 확인 방식 변경 (기존 헤더의 Authorization token을 해체하여 비교하던 방식에서 Authentication 객체를 활용해 로그인한 유저의 email(username)을 확인하는 방식으로 변경) (적용: Post, Comment, Question, Answer, Member)

+++ server/BocamDogam/src/main/java/com/codestates/BocamDogam/answer/service/AnswerServiceImpl.java
@@ -27,14 +27,17 @@ public class AnswerServiceImpl implements AnswerService {
     }
 
     @Override
-    public Answer updateAnswer(Answer answer) {
+    public Answer updateAnswer(Answer answer, String email) {
 
         Answer findAnswer = findVerifiedAnswer(answer.getAnswerId());
+        if(email.compareTo(findAnswer.getMember().getEmail())==0) {
+            Optional.ofNullable(answer.getContent())
+                    .ifPresent(content -> findAnswer.setContent(content));
 
-        Optional.ofNullable(answer.getContent())
-                .ifPresent(content -> findAnswer.setContent(content));
-
-        return answerRepository.save(findAnswer);
+            return answerRepository.save(findAnswer);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_ALLOWED);
+        }
     }
 
     @Override
@@ -50,9 +53,13 @@ public class AnswerServiceImpl implements AnswerService {
     }
 
     @Override
-    public void deleteAnswer(Long answerId) {
+    public void deleteAnswer(Long answerId, String email) {
         Answer findAnswer = findVerifiedAnswer(answerId);
-        answerRepository.delete(findAnswer);
+        if(email.compareTo(findAnswer.getMember().getEmail())==0) {
+            answerRepository.delete(findAnswer);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_ALLOWED);
+        }
     }
 
     @Override
