+++ server/BocamDogam/src/main/java/com/codestates/BocamDogam/post/service/PostServiceImpl.java
@@ -1,21 +1,21 @@
 package com.codestates.BocamDogam.post.service;
 
-import com.codestates.BocamDogam.auth.userdetails.MemberDetailsService;
 import com.codestates.BocamDogam.exception.BusinessLogicException;
 import com.codestates.BocamDogam.exception.ExceptionCode;
+import com.codestates.BocamDogam.member.entity.Member;
 import com.codestates.BocamDogam.post.entity.Post;
 import com.codestates.BocamDogam.post.repository.PostRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
@@ -29,18 +29,22 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public Post updatePost(Post post) {
-
-        Post findPost = findVerifiedPost(post.getPostId());
-
-        Optional.ofNullable(post.getBoard())
-                .ifPresent(board -> findPost.setBoard(board));
-        Optional.ofNullable(post.getTitle())
-                .ifPresent(title -> findPost.setTitle(title));
-        Optional.ofNullable(post.getContent())
-                .ifPresent(content -> findPost.setContent(content));
-
-        return postRepository.save(findPost);
+    public Post updatePost(Post post, String email) {
+        Post findPost = findPost(post.getPostId());
+        if(email.compareTo(findPost.getMember().getEmail()) == 0) {
+            Optional.ofNullable(post.getBoard())
+                    .ifPresent(board -> findPost.setBoard(board));
+            Optional.ofNullable(post.getTitle())
+                    .ifPresent(title -> findPost.setTitle(title));
+            Optional.ofNullable(post.getContent())
+                    .ifPresent(content -> findPost.setContent(content));
+            return postRepository.save(findPost);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_ALLOWED);
+        }
+        // email == findPost.getMember().getEmail() 라고 if문을 작성했을 때 else문으로 넘어갔다. -> email값과 find~ 값을 찍었을 때 콘솔 상 보이는 유저 이메일은 동일했다.
+        // .compareTo() 메서드는 두 문자열을 유니코드로 비교해 동일할 경우 이꼬르 0이라고 하여 A.compareTo(B)==0으로 if문을 수정했더니 기대했던 결과가 나왔다.
+        // 문자열은 == 로 비교하면 안 되는 건지 확인이 필요하다.
     }
 
     @Override
@@ -68,9 +72,13 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, String email) {
         Post findPost = findVerifiedPost(postId);
-        postRepository.delete(findPost);
+        if(email.compareTo(findPost.getMember().getEmail()) == 0) {
+            postRepository.delete(findPost);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_ALLOWED);
+        }
     }
 
     @Override
